pyproject.toml
[project]
name = "voyage-geo"
version = "0.1.0"
description = "Open source Generative Engine Optimization (GEO) analysis tool"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "openai>=1.60.0",
    "anthropic>=0.42.0",
    "google-genai>=1.0.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "vaderSentiment>=3.3.2",
    "pandas>=2.2.0",
    "plotly>=6.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=24.1.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pandas-stubs>=2.2.0",
]

[project.scripts]
voyage-geo = "voyage_geo.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voyage_geo"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 150
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = false
